refactor(util)

+++ server/src/lib/utils/fs.ts
@@ -3,16 +3,27 @@ import * as fs from "node:fs";
 
 /** search to files and deleting file exist */
 export function deleteExistingFile(baseDir: string, fileNameWithoutExt: string) {
-  const files = fs.readdirSync(baseDir);
+  let files: fs.Dirent[] = [];
 
-  for (const file of files) {
+  try {
+    files = fs.readdirSync(baseDir, {withFileTypes: true});
+  } catch (e) {
+    console.error(`Read Dir Error in ${deleteExistingFile.name}`, e);
+  }
+
+  for (const dirent of files) {
+    if (!dirent.isFile()) continue;
+
+    const file: string = dirent.name;
+
     if (file.startsWith(fileNameWithoutExt + ".")) {
-      const fullPath: string = path.join(baseDir, file);
+      const fullPath = path.join(baseDir, file);
 
       try {
         fs.unlinkSync(fullPath);
+        return;
       } catch (e) {
-        console.error("Delete File Error: ", e);
+        console.error(`Error while delete "${fullPath}"`, e);
       }
     }
   }
